Fixing clean code and a little solid.
