+++ Kaizen/Kaizen.Server/Application/Interfaces/Payroll/ISalaryCalculator.cs
@@ -1,4 +1,4 @@
-using Kaizen.Server.Application.Services.Payroll;
+using Kaizen.Server.Application.Dtos.Payroll;
 
 namespace Kaizen.Server.Application.Interfaces.Payroll
 {
