+++ Kaizen/Kaizen.Server/Application/Services/Payroll/PayrollSummaryCalculator.cs
@@ -1,4 +1,5 @@
 using Kaizen.Server.API.Controllers;
+using Kaizen.Server.Application.Dtos.Payroll;
 using Kaizen.Server.Application.Interfaces.Payroll;
 
 namespace Kaizen.Server.Application.Services.Payroll
